requirements.txt
wheel
pyyaml
requests
setuptools 